requirements/requirements-core.txt
# Packages required for core functionality 
numpy 

# JAX optimisation routines
jaxopt

# Sister packages
s2wav >= 1.0.1
s2fft >= 1.1.0